Add files via upload
Please, implement methods in Direction class:


ofDegrees - Return a Direction instance by input degrees or null if there is none.

closestToDegrees - Return the closest Direction instance to input degrees.

opposite - Return a Direction instance that is opposite to this.

differenceDegreesTo - Return the difference in degrees between this and input Direction instance.

Note that degrees input value may be out of the [0; 360) range. Adjust them accordingly.